Move utility functions to utils/contracts.py
Better separation and less danger of circular dependencies
