+++ scenario_player/runner.py
@@ -15,21 +15,15 @@ from raiden_contracts.constants import (
     CHAINNAME_TO_ID,
     CONTRACT_CUSTOM_TOKEN,
     CONTRACT_TOKEN_NETWORK_REGISTRY,
-    CONTRACT_USER_DEPOSIT,
 )
-from raiden_contracts.contract_manager import (
-    ContractManager,
-    DeployedContracts,
-    contracts_precompiled_path,
-    get_contracts_deployment_info,
-)
+from raiden_contracts.contract_manager import DeployedContracts, get_contracts_deployment_info
 from requests import HTTPError, Session
 from web3 import HTTPProvider, Web3
 
 from raiden.accounts import Account
 from raiden.constants import UINT256_MAX
 from raiden.network.proxies.custom_token import CustomToken
-from raiden.network.proxies.proxy_manager import ProxyManager, ProxyManagerMetadata
+from raiden.network.proxies.proxy_manager import ProxyManager
 from raiden.network.proxies.token_network_registry import TokenNetworkRegistry
 from raiden.network.proxies.user_deposit import UserDeposit
 from raiden.network.rpc.client import JSONRPCClient
@@ -38,13 +32,11 @@ from raiden.utils.formatting import to_canonical_address
 from raiden.utils.nursery import Janitor
 from raiden.utils.typing import (
     Address,
-    BlockNumber,
     ChainID,
     TokenAddress,
     TokenAmount,
     TokenNetworkAddress,
     TokenNetworkRegistryAddress,
-    UserDepositAddress,
 )
 from scenario_player.constants import (
     API_URL_TOKEN_NETWORK_ADDRESS,
@@ -60,6 +52,10 @@ from scenario_player.exceptions.legacy import TokenNetworkDiscoveryTimeout
 from scenario_player.node_support import NodeController, RaidenReleaseKeeper
 from scenario_player.utils import TimeOutHTTPAdapter
 from scenario_player.utils.configuration.settings import SettingsConfig, UDCSettingsConfig
+from scenario_player.utils.contracts import (
+    get_proxy_manager,
+    get_udc_and_corresponding_token_from_dependencies,
+)
 from scenario_player.utils.token import (
     TokenDetails,
     eth_maybe_transfer,
@@ -99,53 +95,6 @@ def wait_for_nodes_to_be_ready(node_runners, session):
                 except requests.exceptions.RequestException:
                     pass
                 gevent.sleep(0.5)
-
-
-def get_proxy_manager(client: JSONRPCClient, deploy: DeployedContracts) -> ProxyManager:
-    contract_manager = ContractManager(contracts_precompiled_path(RAIDEN_CONTRACT_VERSION))
-
-    assert "contracts" in deploy, deploy
-    token_network_deployment_details = deploy["contracts"][CONTRACT_TOKEN_NETWORK_REGISTRY]
-    deployed_at = token_network_deployment_details["block_number"]
-    token_network_registry_deployed_at = BlockNumber(deployed_at)
-
-    return ProxyManager(
-        client,
-        contract_manager,
-        ProxyManagerMetadata(
-            token_network_registry_deployed_at=token_network_registry_deployed_at,
-            filters_start_at=token_network_registry_deployed_at,
-        ),
-    )
-
-
-def get_udc_and_corresponding_token_from_dependencies(
-    udc_address: Optional[ChecksumAddress], chain_id: ChainID, proxy_manager: ProxyManager
-) -> Tuple[UserDeposit, CustomToken]:
-    """ Return contract proxies for the UserDepositContract and associated token.
-
-    This will return a proxy to the `UserDeposit` contract as determined by the
-    **local** Raiden depedency.
-    """
-    if udc_address is None:
-
-        contracts = get_contracts_deployment_info(chain_id, version=RAIDEN_CONTRACT_VERSION)
-
-        msg = (
-            f"invalid chain_id, {chain_id} is not available for version {RAIDEN_CONTRACT_VERSION}"
-        )
-        assert contracts, msg
-
-        udc_address = contracts["contracts"][CONTRACT_USER_DEPOSIT]["address"]
-
-    userdeposit_proxy = proxy_manager.user_deposit(
-        UserDepositAddress(to_canonical_address(udc_address)), "latest"
-    )
-
-    token_address = userdeposit_proxy.token_address("latest")
-    user_token_proxy = proxy_manager.custom_token(token_address, "latest")
-
-    return userdeposit_proxy, user_token_proxy
 
 
 def get_token_network_registry_from_dependencies(
